requirements.txt
requests>=2.25.0
beautifulsoup4>=4.9.0
urllib3>=1.26.0
aiohttp>=3.8.0
python-dotenv>=0.19.0

